+++ src/Event/Genesis_Events.php
@@ -15,45 +15,15 @@ class Genesis_Events implements Hook_Event_Config_Interface {
 	 */
 	public function get_events(): array {
 
-		$events = [];
-
-		if ( ! \function_exists( 'genesis' ) ) {
-			return $events;
-		}
-
-		$events['genesis_before_header'][10] = [
-			'event'    => 'Header',
-			'category' => 'theme',
-		];
-
-		$events['genesis_after_header'][10] = [
-			'action'   => 'stop',
-			'event'    => 'Header',
-			'category' => 'theme',
-		];
-
-		$events['genesis_before_sidebar_widget_area'][10] = [
-			'event'    => 'Sidebar',
-			'category' => 'theme',
-		];
-
-		$events['genesis_after_sidebar_widget_area'][10] = [
-			'action'   => 'stop',
-			'event'    => 'Sidebar',
-			'category' => 'theme',
-		];
-
-		$events['genesis_before_sidebar_alt_widget_area'][10] = [
-			'event'    => 'Sidebar (alternate)',
-			'category' => 'theme',
-		];
-
-		$events['genesis_after_sidebar_alt_widget_area'][10] = [
-			'action'   => 'stop',
-			'event'    => 'Sidebar (alternate)',
-			'category' => 'theme',
-		];
-
-		return $events;
+		return \function_exists( 'genesis' ) ? [
+			[ 'Header', 'theme', 'genesis_before_header', 'genesis_after_header' ],
+			[ 'Sidebar', 'theme', 'genesis_before_sidebar_widget_area', 'genesis_after_sidebar_widget_area' ],
+			[
+				'Sidebar (alternate)',
+				'theme',
+				'genesis_before_sidebar_alt_widget_area',
+				'genesis_after_sidebar_alt_widget_area',
+			],
+		] : [];
 	}
 }