+++ src/Event/THA_Events.php
@@ -15,34 +15,9 @@ class THA_Events implements Hook_Event_Config_Interface {
 	 */
 	public function get_events(): array {
 
-		$events = [];
-
-		if ( ! \defined( 'THA_HOOKS_VERSION' ) ) {
-			return $events;
-		}
-
-		$events['tha_header_before'][10] = [
-			'event'    => 'Header',
-			'category' => 'theme',
-		];
-
-		$events['tha_header_after'][10] = [
-			'action'   => 'stop',
-			'event'    => 'Header',
-			'category' => 'theme',
-		];
-
-		$events['tha_sidebars_before'][10] = [
-			'event'    => 'Sidebars',
-			'category' => 'theme',
-		];
-
-		$events['tha_sidebars_after'][10] = [
-			'action'   => 'stop',
-			'event'    => 'Sidebars',
-			'category' => 'theme',
-		];
-
-		return $events;
+		return \defined( 'THA_HOOKS_VERSION' ) ? [
+			[ 'Header', 'theme', 'tha_header_before', 'tha_header_after' ],
+			[ 'Sidebars', 'theme', 'tha_sidebars_before', 'tha_sidebars_after' ],
+		] : [];
 	}
 }